Remove useTestConnection to reduce depth of nested hooks
